progress

+++ lista04/bst.h
@@ -12,6 +12,6 @@ Node* min_value_node(Node *node);
 void free_tree(Node *root);
 size_t get_size(void);
 
-void print_bst(const Node* restrict root, const size_t depth, const char prefix);
+void print_bst(const Node* restrict root, const int32_t depth, const char prefix);
 void init_traces(const size_t size);
 void free_traces(void);
